bug fixes
https://app.asana.com/1/1138832104141584/task/1211495175765832

+++ includes/blocks/ac-profile-picture/init.php
@@ -182,24 +182,36 @@ class init extends Blocks
             return false;
         }
 
-        // User ID
-        $user_id = sanitize_text_field(wp_unslash($form['user_id']));
+        $user_id = get_current_user_id();
+        $user = $user_id ? get_user_by('id', $user_id) : null;
+        $person_uuid = $user instanceof \WP_User ? $user->user_login : null;
+        $primary_identifier = $person_uuid ?: (string) $user_id;
+        if (empty($primary_identifier)) {
+            $this->setError('missing_user', __('Could not determine the current user for upload.', 'wicket-acc'));
 
-        // Remove any existing file on wicket-profile-pictures/{user_id}.{extension}
-        $file_path = $this->pp_uploads_path . $user_id . '.' . $file_extension;
+            return false;
+        }
 
-        // Delete the file if it exists
-        foreach ($this->pp_extensions as $ext) {
-            $other_file_path = $this->pp_uploads_path . $user_id . '.' . $ext;
+        // Remove any existing file on wicket-profile-pictures/{uuid}.{extension}
+        $file_path = $this->pp_uploads_path . $primary_identifier . '.' . $file_extension;
 
-            if (file_exists($other_file_path)) {
-                wp_delete_file($other_file_path);
+        $identifiers = array_filter(array_unique([
+            $person_uuid,
+            (string) $user_id,
+        ]));
+        foreach ($identifiers as $identifier) {
+            foreach ($this->pp_extensions as $ext) {
+                $other_file_path = $this->pp_uploads_path . $identifier . '.' . $ext;
+
+                if (file_exists($other_file_path)) {
+                    wp_delete_file($other_file_path);
+                }
             }
         }
 
-        // No matter whats the file name, rename it to {user_id}.{extension}
-        $_FILES['profile-image']['name'] = $user_id . '.' . $file_extension;
-        $_FILES['profile-image']['full_path'] = $user_id . '.' . $file_extension;
+        // No matter whats the file name, rename it to {uuid}.{extension}
+        $_FILES['profile-image']['name'] = $primary_identifier . '.' . $file_extension;
+        $_FILES['profile-image']['full_path'] = $primary_identifier . '.' . $file_extension;
 
         // Create subfolder if it doesn't exist
         if (!file_exists($this->pp_uploads_path)) {
@@ -245,6 +257,8 @@ class init extends Blocks
              * @var string|null $profile_image_url URL of the updated profile image, or null if not set.
              */
             do_action('wicket/acc/profile/edit/profile_image_updated', $profile_photo_url);
+
+            WACC()->Profile()->syncProfileImageToMdp($profile_photo_url);
         }
 
         return true;
@@ -275,15 +289,22 @@ class init extends Blocks
             return false;
         }
 
-        // User ID
-        $user_id = absint($form['user_id']);
+        $user_id = get_current_user_id();
+        $user = $user_id ? get_user_by('id', $user_id) : null;
+        $person_uuid = $user instanceof \WP_User ? $user->user_login : null;
 
-        // Remove any existing file on wicket-profile-pictures/{user_id}.{extension}
-        foreach ($this->pp_extensions as $ext) {
-            $file_path = $this->pp_uploads_path . $user_id . '.' . $ext;
+        // Remove any existing file on wicket-profile-pictures/{uuid}.{extension}
+        $identifiers = array_filter(array_unique([
+            $person_uuid,
+            (string) $user_id,
+        ]));
+        foreach ($identifiers as $identifier) {
+            foreach ($this->pp_extensions as $ext) {
+                $file_path = $this->pp_uploads_path . $identifier . '.' . $ext;
 
-            if (file_exists($file_path)) {
-                wp_delete_file($file_path);
+                if (file_exists($file_path)) {
+                    wp_delete_file($file_path);
+                }
             }
         }
 
@@ -291,6 +312,7 @@ class init extends Blocks
          * @var string|null $profile_image_url URL of the updated profile image, or null if not set.
          */
         do_action('wicket/acc/profile/edit/profile_image_updated', null);
+        WACC()->Profile()->syncProfileImageToMdp(null);
 
         return true;
     }
